Cache pet request collections, as well as data.
I'm noticing a minor perf issue on the index page. My guess is that the process of sorting the data and figuring out the cache key isn't blazing-fast like I'd hope.

So, now, we load the data very quickly in the controller, and Rails very quickly figures out the cache key for the view (we use their native stuff now). We could conceivably remove that middle step, but it seems like extra work and extra coupling, so, shrug. Hope this is good enough!