Add explicit success logging to Discord job
Something stopped working here. I thiiink it's not a regression on our end? And this logging seems to be further confirming that, because I see the success message (indicating 200 OK) when trying to reproduce the issue from dev?

Anyway, this is still helpful logging to have, so I'm adding it!